mcp/requirements.txt
mcp[cli]
python-dotenv
pydantic
asyncpg
openai